Enum delete_table::Error

#[non_exhaustive]
pub(crate) enum Error {

    BackupInUseException(BackupInUseException),
    BackupNotFoundException(BackupNotFoundException),
    ConditionalCheckFailedException(ConditionalCheckFailedException),
    ContinuousBackupsUnavailableException(ContinuousBackupsUnavailableException),
    DuplicateItemException(DuplicateItemException),
    ExportConflictException(ExportConflictException),
    ExportNotFoundException(ExportNotFoundException),
    GlobalTableAlreadyExistsException(GlobalTableAlreadyExistsException),
    GlobalTableNotFoundException(GlobalTableNotFoundException),
    IdempotentParameterMismatchException(IdempotentParameterMismatchException),
    IndexNotFoundException(IndexNotFoundException),
    InternalServerError(InternalServerError),
    InvalidEndpointException(InvalidEndpointException),
    InvalidExportTimeException(InvalidExportTimeException),
    InvalidRestoreTimeException(InvalidRestoreTimeException),
    ItemCollectionSizeLimitExceededException(ItemCollectionSizeLimitExceededException),
    LimitExceededException(LimitExceededException),
    PointInTimeRecoveryUnavailableException(PointInTimeRecoveryUnavailableException),
    ProvisionedThroughputExceededException(ProvisionedThroughputExceededException),
    ReplicaAlreadyExistsException(ReplicaAlreadyExistsException),
    ReplicaNotFoundException(ReplicaNotFoundException),
    RequestLimitExceeded(RequestLimitExceeded),
    ResourceInUseException(ResourceInUseException),
    ResourceNotFoundException(ResourceNotFoundException),
    TableAlreadyExistsException(TableAlreadyExistsException),
    TableInUseException(TableInUseException),
    TableNotFoundException(TableNotFoundException),
    TransactionCanceledException(TransactionCanceledException),
    TransactionConflictException(TransactionConflictException),
    TransactionInProgressException(TransactionInProgressException),
    Unhandled(Box<dyn Error + Send + Sync + 'static, Global>),

}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

BackupInUseException(BackupInUseException)

There is another ongoing conflicting backup control plane operation on the table. The backup is either being created, deleted or restored to a table.

Tuple Fields of BackupInUseException

0: BackupInUseException

BackupNotFoundException(BackupNotFoundException)

Backup not found for the given BackupARN.

Tuple Fields of BackupNotFoundException

0: BackupNotFoundException

ConditionalCheckFailedException(ConditionalCheckFailedException)

A condition specified in the operation could not be evaluated.

Tuple Fields of ConditionalCheckFailedException

0: ConditionalCheckFailedException

ContinuousBackupsUnavailableException(ContinuousBackupsUnavailableException)

Backups have not yet been enabled for this table.

Tuple Fields of ContinuousBackupsUnavailableException

0: ContinuousBackupsUnavailableException

DuplicateItemException(DuplicateItemException)

There was an attempt to insert an item with the same primary key as an item that already exists in the DynamoDB table.

Tuple Fields of DuplicateItemException

0: DuplicateItemException

ExportConflictException(ExportConflictException)

There was a conflict when writing to the specified S3 bucket.

Tuple Fields of ExportConflictException

0: ExportConflictException

ExportNotFoundException(ExportNotFoundException)

The specified export was not found.

Tuple Fields of ExportNotFoundException

0: ExportNotFoundException

GlobalTableAlreadyExistsException(GlobalTableAlreadyExistsException)

The specified global table already exists.

Tuple Fields of GlobalTableAlreadyExistsException

0: GlobalTableAlreadyExistsException

GlobalTableNotFoundException(GlobalTableNotFoundException)

The specified global table does not exist.

Tuple Fields of GlobalTableNotFoundException

0: GlobalTableNotFoundException

IdempotentParameterMismatchException(IdempotentParameterMismatchException)

DynamoDB rejected the request because you retried a request with a different payload but with an idempotent token that was already used.

Tuple Fields of IdempotentParameterMismatchException

0: IdempotentParameterMismatchException

IndexNotFoundException(IndexNotFoundException)

The operation tried to access a nonexistent index.

Tuple Fields of IndexNotFoundException

0: IndexNotFoundException

InternalServerError(InternalServerError)

An error occurred on the server side.

Tuple Fields of InternalServerError

0: InternalServerError

InvalidEndpointException(InvalidEndpointException)

Tuple Fields of InvalidEndpointException

0: InvalidEndpointException

InvalidExportTimeException(InvalidExportTimeException)

The specified ExportTime is outside of the point in time recovery window.

Tuple Fields of InvalidExportTimeException

0: InvalidExportTimeException

InvalidRestoreTimeException(InvalidRestoreTimeException)

An invalid restore time was specified. RestoreDateTime must be between EarliestRestorableDateTime and LatestRestorableDateTime.

Tuple Fields of InvalidRestoreTimeException

0: InvalidRestoreTimeException

ItemCollectionSizeLimitExceededException(ItemCollectionSizeLimitExceededException)

An item collection is too large. This exception is only returned for tables that have one or more local secondary indexes.

Tuple Fields of ItemCollectionSizeLimitExceededException

0: ItemCollectionSizeLimitExceededException

LimitExceededException(LimitExceededException)

There is no limit to the number of daily on-demand backups that can be taken.

Up to 50 simultaneous table operations are allowed per account. These operations include CreateTable, UpdateTable, DeleteTable,UpdateTimeToLive, RestoreTableFromBackup, and RestoreTableToPointInTime.

The only exception is when you are creating a table with one or more secondary indexes. You can have up to 25 such requests running at a time; however, if the table or index specifications are complex, DynamoDB might temporarily reduce the number of concurrent operations.

There is a soft account quota of 256 tables.

Tuple Fields of LimitExceededException

0: LimitExceededException

PointInTimeRecoveryUnavailableException(PointInTimeRecoveryUnavailableException)

Point in time recovery has not yet been enabled for this source table.

Tuple Fields of PointInTimeRecoveryUnavailableException

0: PointInTimeRecoveryUnavailableException

ProvisionedThroughputExceededException(ProvisionedThroughputExceededException)

Your request rate is too high. The AWS SDKs for DynamoDB automatically retry requests that receive this exception. Your request is eventually successful, unless your retry queue is too large to finish. Reduce the frequency of requests and use exponential backoff. For more information, go to Error Retries and Exponential Backoff in the Amazon DynamoDB Developer Guide.

Tuple Fields of ProvisionedThroughputExceededException

0: ProvisionedThroughputExceededException

ReplicaAlreadyExistsException(ReplicaAlreadyExistsException)

The specified replica is already part of the global table.

Tuple Fields of ReplicaAlreadyExistsException

0: ReplicaAlreadyExistsException

ReplicaNotFoundException(ReplicaNotFoundException)

The specified replica is no longer part of the global table.

Tuple Fields of ReplicaNotFoundException

0: ReplicaNotFoundException

RequestLimitExceeded(RequestLimitExceeded)

Throughput exceeds the current throughput quota for your account. Please contact AWS Support at AWS Support to request a quota increase.

Tuple Fields of RequestLimitExceeded

0: RequestLimitExceeded

ResourceInUseException(ResourceInUseException)

The operation conflicts with the resource's availability. For example, you attempted to recreate an existing table, or tried to delete a table currently in the CREATING state.

Tuple Fields of ResourceInUseException

0: ResourceInUseException

ResourceNotFoundException(ResourceNotFoundException)

The operation tried to access a nonexistent table or index. The resource might not be specified correctly, or its status might not be ACTIVE.

Tuple Fields of ResourceNotFoundException

0: ResourceNotFoundException

TableAlreadyExistsException(TableAlreadyExistsException)

A target table with the specified name already exists.

Tuple Fields of TableAlreadyExistsException

0: TableAlreadyExistsException

TableInUseException(TableInUseException)

A target table with the specified name is either being created or deleted.

Tuple Fields of TableInUseException

0: TableInUseException

TableNotFoundException(TableNotFoundException)

A source table with the name TableName does not currently exist within the subscriber's account.

Tuple Fields of TableNotFoundException

0: TableNotFoundException

TransactionCanceledException(TransactionCanceledException)

The entire transaction request was canceled.

DynamoDB cancels a TransactWriteItems request under the following circumstances:

A condition in one of the condition expressions is not met.

A table in the TransactWriteItems request is in a different account or region.

More than one action in the TransactWriteItems operation targets the same item.

There is insufficient provisioned capacity for the transaction to be completed.

An item size becomes too large (larger than 400 KB), or a local secondary index (LSI) becomes too large, or a similar validation error occurs because of changes made by the transaction.

There is a user error, such as an invalid data format.

DynamoDB cancels a TransactGetItems request under the following circumstances:

There is an ongoing TransactGetItems operation that conflicts with a concurrent PutItem, UpdateItem, DeleteItem or TransactWriteItems request. In this case the TransactGetItems operation fails with a TransactionCanceledException.

A table in the TransactGetItems request is in a different account or region.

There is insufficient provisioned capacity for the transaction to be completed.

There is a user error, such as an invalid data format.

If using Java, DynamoDB lists the cancellation reasons on the CancellationReasons property. This property is not set for other languages. Transaction cancellation reasons are ordered in the order of requested items, if an item has no error it will have NONE code and Null message.

Cancellation reason codes and possible error messages:

No Errors:

Code: NONE

Message: null

Conditional Check Failed:

Code: ConditionalCheckFailed

Message: The conditional request failed.

Item Collection Size Limit Exceeded:

Code: ItemCollectionSizeLimitExceeded

Message: Collection size exceeded.

Transaction Conflict:

Code: TransactionConflict

Message: Transaction is ongoing for the item.

Provisioned Throughput Exceeded:

Code: ProvisionedThroughputExceeded

Messages:

The level of configured provisioned throughput for the table was exceeded. Consider increasing your provisioning level with the UpdateTable API.

This Message is received when provisioned throughput is exceeded is on a provisioned DynamoDB table.

The level of configured provisioned throughput for one or more global secondary indexes of the table was exceeded. Consider increasing your provisioning level for the under-provisioned global secondary indexes with the UpdateTable API.

This message is returned when provisioned throughput is exceeded is on a provisioned GSI.

Throttling Error:

Code: ThrottlingError

Messages:

Throughput exceeds the current capacity of your table or index. DynamoDB is automatically scaling your table or index so please try again shortly. If exceptions persist, check if you have a hot key: https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/bp-partition-key-design.html.

This message is returned when writes get throttled on an On-Demand table as DynamoDB is automatically scaling the table.

Throughput exceeds the current capacity for one or more global secondary indexes. DynamoDB is automatically scaling your index so please try again shortly.

This message is returned when when writes get throttled on an On-Demand GSI as DynamoDB is automatically scaling the GSI.

Validation Error:

Code: ValidationError

Messages:

One or more parameter values were invalid.

The update expression attempted to update the secondary index key beyond allowed size limits.

The update expression attempted to update the secondary index key to unsupported type.

An operand in the update expression has an incorrect data type.

Item size to update has exceeded the maximum allowed size.

Number overflow. Attempting to store a number with magnitude larger than supported range.

Type mismatch for attribute to update.

Nesting Levels have exceeded supported limits.

The document path provided in the update expression is invalid for update.

The provided expression refers to an attribute that does not exist in the item.

Tuple Fields of TransactionCanceledException

0: TransactionCanceledException

TransactionConflictException(TransactionConflictException)

Operation was rejected because there is an ongoing transaction for the item.

Tuple Fields of TransactionConflictException

0: TransactionConflictException

TransactionInProgressException(TransactionInProgressException)

The transaction with the given request token is already in progress.

Tuple Fields of TransactionInProgressException

0: TransactionInProgressException

Unhandled(Box<dyn Error + Send + Sync + 'static, Global>)

An unhandled error occurred.

Tuple Fields of Unhandled

0: Box<dyn Error + Send + Sync + 'static, Global>